+++ WalkingDeadCharacters/Controllers/SeasonController.cs
@@ -33,6 +33,9 @@ namespace WalkingDeadCharacters.Controllers
 
             var service = CreateSeasonService();
 
+            if (season == default)
+                return NotFound();
+
             if (!service.CreateSeason(season))
                 return InternalServerError();
 
@@ -43,17 +46,22 @@ namespace WalkingDeadCharacters.Controllers
         {
             SeasonService seasonService = CreateSeasonService();
             var season = seasonService.GetSeasonById(id);
+
+            if (season == default)
+            {
+                return NotFound();
+            }
             return Ok(season);
         }
 
-        public IHttpActionResult Put(SeasonEdit season)
+        public IHttpActionResult Put(SeasonEdit season, int id)
         {
             if (!ModelState.IsValid)
                 return BadRequest(ModelState);
 
             var service = CreateSeasonService();
 
-            if (!service.UpdateSeason(season))
+            if (!service.UpdateSeason(season, id))
                 return InternalServerError();
 
             return Ok();
@@ -62,6 +70,13 @@ namespace WalkingDeadCharacters.Controllers
         public IHttpActionResult Delete(int id)
         {
             var service = CreateSeasonService();
+
+            SeasonDetail deleteSeason = service.GetSeasonById(id);
+
+            if (deleteSeason == default)
+            {
+                return NotFound();
+            }
 
             if (!service.DeleteSeason(id))
                 return InternalServerError();
